+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Each configured site domain now **requires** you to assign an available ip address
 * You can now define the uid a node site runs as
 * Fix `Site#matches` so it also searches through defined `hostname`s
+* `Site#getAddress` will callback with an error when it fails to start
 
 ## 0.1.0